PI_camera/requirements.txt
opencv-python
onnxruntime
paho-mqtt<2.0.0
numpy
